+++ src/pages/Detail/ProductDescription/ProductDescription.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Map } from 'react-kakao-maps-sdk';
+import { Map, MapMarker } from 'react-kakao-maps-sdk';
 import styled from 'styled-components';
 
 const ProductDescription = ({ productDetailImages, latitude, longitude }) => {
@@ -16,9 +16,11 @@ const ProductDescription = ({ productDetailImages, latitude, longitude }) => {
       </Description>
       <Map
         center={{ lat: latitude, lng: longitude }} // 지도의 중심 좌표
-        style={{ width: '100%', height: '300px' }} // 지도 크기
+        style={{ width: '100%', height: '32vw' }} // 지도 크기
         level={3} // 지도 확대 레벨
-      />
+      >
+        <MapMarker position={{ lat: latitude, lng: longitude }} />
+      </Map>
     </>
   );
 };
